Add additional check to build.cmake
Add the check for terarepo/src. From a fresh git checkout, the terarepo/src directory doesn’t exist (it needs a submodule update) and isn’t detected until ~80% compiled. So check for the directory at the beginning of the script and alert user if not present. Additionally, remove building python in boost. It’s not used by Vaa3D, speeds up compilation, and prevents errors if the development headers are not installed.