Update NeuronAnnotator to handle H5J files
To save space, the files NA requires are sometimes purged and replaced with H5Js. This change allows the NA progressive loader to fallback to an H5J if the other files are missing. This required adding guards around access to the three NA volume stacks as they all may not be available in H5J mode.