Staged Loader now looks for best candidate of each type
Sometimes, we can have partially purged results. The Neuron Annotator would take the results from the first viable sample it finds (typically the reference channel), and applies that to the label and signal channels. This commit changes this so that each of the channels is checked independently for the best sample to use.